Arrumando classe abstrata e hierarquias
