+++ Trabalhos/src/negocio/Filme.java
@@ -1,6 +1,6 @@
-package classes;
+package negocio;
 
-import java.util.Date;
+import java.util.*;
 
 public class Filme {
