+++ Trabalhos/src/negocio/Sessao.java
@@ -1,4 +1,4 @@
-package classes;
+package negocio;
 
 public class Sessao {
